test: 저장된 게임의 존재여부 확인 추가
저장된 게임의 존재여부를 확인하는 코드를 추가하여 테스트 정확도를 높일 수 있도록 하였습니다.
